+++ packages/web3-eth-contract/src/constant.ts
@@ -14,19 +14,4 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
-import { AbiEventFragment, LogsInput } from 'web3-types';
-import { decodeEventABI } from '../../src/encoding';
-import { decodeEventABIData } from '../fixtures/encoding';
-
-describe('encoding decoding functions', () => {
-	describe('decode', () => {
-		describe('decodeEventABI', () => {
-			it.each(decodeEventABIData)(
-				'%s',
-				(event: AbiEventFragment & { signature: string }, inputs: LogsInput, output) => {
-					expect(decodeEventABI(event, inputs, [])).toBe(output);
-				},
-			);
-		});
-	});
-});
+export { ALL_EVENTS, ALL_EVENTS_ABI } from 'web3-eth';