+++ packages/web3-eth/test/fixtures/erc20.ts
@@ -1,0 +1,175 @@
+﻿/*
+This file is part of web3.js.
+
+web3.js is free software: you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+web3.js is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public License
+along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
+*/
+export const ERC20TokenAbi = [
+	{
+		inputs: [{ internalType: 'uint256', name: 'initialSupply', type: 'uint256' }],
+		stateMutability: 'nonpayable',
+		type: 'constructor',
+		signature: '',
+	},
+	{
+		anonymous: false,
+		inputs: [
+			{ indexed: true, internalType: 'address', name: 'owner', type: 'address' },
+			{ indexed: true, internalType: 'address', name: 'spender', type: 'address' },
+			{ indexed: false, internalType: 'uint256', name: 'value', type: 'uint256' },
+		],
+		name: 'Approval',
+		type: 'event',
+		signature: '0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925',
+	},
+	{
+		anonymous: false,
+		inputs: [
+			{ indexed: true, internalType: 'address', name: 'from', type: 'address' },
+			{ indexed: true, internalType: 'address', name: 'to', type: 'address' },
+			{ indexed: false, internalType: 'uint256', name: 'value', type: 'uint256' },
+		],
+		name: 'Transfer',
+		type: 'event',
+		signature: '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
+	},
+	{
+		inputs: [
+			{ internalType: 'address', name: 'owner', type: 'address' },
+			{ internalType: 'address', name: 'spender', type: 'address' },
+		],
+		name: 'allowance',
+		outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
+		stateMutability: 'view',
+		type: 'function',
+		signature: '0xdd62ed3e',
+		constant: true,
+		payable: false,
+	},
+	{
+		inputs: [
+			{ internalType: 'address', name: 'spender', type: 'address' },
+			{ internalType: 'uint256', name: 'amount', type: 'uint256' },
+		],
+		name: 'approve',
+		outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
+		stateMutability: 'nonpayable',
+		type: 'function',
+		signature: '0x095ea7b3',
+		constant: false,
+		payable: false,
+	},
+	{
+		inputs: [{ internalType: 'address', name: 'account', type: 'address' }],
+		name: 'balanceOf',
+		outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
+		stateMutability: 'view',
+		type: 'function',
+		signature: '0x70a08231',
+		constant: true,
+		payable: false,
+	},
+	{
+		inputs: [],
+		name: 'decimals',
+		outputs: [{ internalType: 'uint8', name: '', type: 'uint8' }],
+		stateMutability: 'view',
+		type: 'function',
+		signature: '0x313ce567',
+		constant: true,
+		payable: false,
+	},
+	{
+		inputs: [
+			{ internalType: 'address', name: 'spender', type: 'address' },
+			{ internalType: 'uint256', name: 'subtractedValue', type: 'uint256' },
+		],
+		name: 'decreaseAllowance',
+		outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
+		stateMutability: 'nonpayable',
+		type: 'function',
+		signature: '0xa457c2d7',
+		constant: false,
+		payable: false,
+	},
+	{
+		inputs: [
+			{ internalType: 'address', name: 'spender', type: 'address' },
+			{ internalType: 'uint256', name: 'addedValue', type: 'uint256' },
+		],
+		name: 'increaseAllowance',
+		outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
+		stateMutability: 'nonpayable',
+		type: 'function',
+		signature: '0x39509351',
+		constant: false,
+		payable: false,
+	},
+	{
+		inputs: [],
+		name: 'name',
+		outputs: [{ internalType: 'string', name: '', type: 'string' }],
+		stateMutability: 'view',
+		type: 'function',
+		signature: '0x06fdde03',
+		constant: true,
+		payable: false,
+	},
+	{
+		inputs: [],
+		name: 'symbol',
+		outputs: [{ internalType: 'string', name: '', type: 'string' }],
+		stateMutability: 'view',
+		type: 'function',
+		signature: '0x95d89b41',
+		constant: true,
+		payable: false,
+	},
+	{
+		inputs: [],
+		name: 'totalSupply',
+		outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
+		stateMutability: 'view',
+		type: 'function',
+		signature: '0x18160ddd',
+		constant: true,
+		payable: false,
+	},
+	{
+		inputs: [
+			{ internalType: 'address', name: 'to', type: 'address' },
+			{ internalType: 'uint256', name: 'amount', type: 'uint256' },
+		],
+		name: 'transfer',
+		outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
+		stateMutability: 'nonpayable',
+		type: 'function',
+		signature: '0xa9059cbb',
+		constant: false,
+		payable: false,
+	},
+	{
+		inputs: [
+			{ internalType: 'address', name: 'from', type: 'address' },
+			{ internalType: 'address', name: 'to', type: 'address' },
+			{ internalType: 'uint256', name: 'amount', type: 'uint256' },
+		],
+		name: 'transferFrom',
+		outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
+		stateMutability: 'nonpayable',
+		type: 'function',
+		signature: '0x23b872dd',
+		constant: false,
+		payable: false,
+	},
+] as const;