+++ CHANGELOG.md
@@ -2007,7 +2007,6 @@ If there are any bugs, improvements, optimizations or any new feature proposal f
 
 #### web3-eth-contract
 
-
 #### web3-utils
 
 -   `soliditySha3()` with BigInt support
@@ -2021,6 +2020,7 @@ If there are any bugs, improvements, optimizations or any new feature proposal f
 #### web3-eth
 
 -   Added to `Web3Config` property `contractDataInputFill` allowing users to have the choice using property `data`, `input` or `both` for contract methods to be sent to the RPC provider when creating contracts. (#6377) (#6400)
+-   Added `ALL_EVENTS` and `ALL_EVENTS_ABI` constants, `SendTransactionEventsBase` type, `decodeEventABI` method (#6410)
 
 #### web3-eth-contract
 
@@ -2081,5 +2081,26 @@ If there are any bugs, improvements, optimizations or any new feature proposal f
 
 -   Dependencies updated
 
-
 ## [Unreleased]
+
+### Added
+
+#### web3-eth
+
+-   Added `ALL_EVENTS` and `ALL_EVENTS_ABI` constants, `SendTransactionEventsBase` type, `decodeEventABI` method (#6410)
+
+#### web3-types
+
+-   Interface `EventLog` was added. (#6410)
+
+### Fixed
+
+#### web3-eth
+
+-   Ensure provider.supportsSubscriptions exists before watching by subscription (#6440)
+
+### Changed
+
+#### web3-eth-contract
+
+-   The `events` property was added to the `receipt` object (#6410)